curve25519: update package structure per golang.org/wiki/TargetSpecific
This was extracted from CL 205157.

Updates golang/go#31470

Reviewed-on: https://go-review.googlesource.com/c/crypto/+/205861
