+++ curve25519/vectors_test.go
